Update event emitters to Subject
Modified the code to use Subject instead of event emitters. Also removed
some event emitter code that was no longer needed because of routes in
place now. Needed to modify the emit() to next(). No change to the
subscription, other than to add unsubscribe now that it is our own
Subject and not an Angular observable.